// src/app/models/accidents.model.ts
export interface Accidents {
    id?: number;
    profession: string; // Use string for enum in Angular
    capitalDeces: number;
    capitalIPP: number;
    montantRenteParJourDT: number;
    dureeFranchise: string;
    capitalTraitement: number;
    nom: string;
    prenom: string;
    mail: string;
    telephone: string;
  }